requirements.txt
fastapi==0.135.1
moviepy==2.2.1
openai-whisper==20250625
pydantic==2.12.5
python-multipart==0.0.22
torch==2.10.0
torchaudio==2.10.0
torchvision==0.25.0
uvicorn==0.42.0
